Move document-state to services (#6336)
* Move document-state to services

* fixes

* fixes
