+++ build-system/tasks/presubmit-checks.js
@@ -108,6 +108,16 @@ var forbiddenTerms = {
   '\\.prefetch\\(': {
     message: 'Do not use preconnect.prefetch, use preconnect.preload instead.'
   },
+  'documentStateFor': {
+    message: privateServiceFactory,
+    whitelist: [
+      'src/custom-element.js',
+      'src/style-installer.js',
+      'src/service/document-state.js',
+      'src/service/viewer-impl.js',
+      'src/service/vsync-impl.js',
+    ],
+  },
   'iframePing': {
     message: 'This is only available in vendor config for ' +
         'temporary workarounds.',
