+++ extensions/amp-analytics/0.1/test/test-instrumentation.js
@@ -28,7 +28,7 @@ import {installPlatformService} from '../../../../src/service/platform-impl';
 import {
     installResourcesServiceForDoc,
 } from '../../../../src/service/resources-impl';
-import {documentStateFor} from '../../../../src/document-state';
+import {documentStateFor} from '../../../../src/service/document-state';
 
 
 adopt(window);
