+++ test/functional/test-document-state.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {DocumentState} from '../../src/document-state';
+import {DocumentState} from '../../src/service/document-state';
 import * as dom from '../../src/dom';
 import * as sinon from 'sinon';
